refactor: move stopwatch source file
